+++ routes/mainRoute.js
@@ -12,59 +12,9 @@ const excel = require("excel4node");
 const logger = require("../src/logger");
 const config = require("../src/configLoader");
 const httpErrors = require("../resources/httpErrors.json");
+const mainLib = require("../lib/mainLib");
 
 let router = express.Router();
-
-function checkKeys(dataObj, prefix) {
-    prefix = prefix || "";
-    return [...new Set(dataObj.map(val => { 
-        let arr = [];
-        for(let key in val) {
-            if(typeof val[key] === "object") arr.push(checkKeys([val[key]], `${prefix}${key}.`));
-            else arr.push(`${prefix}${key}`);
-        } 
-        return arr;
-    }).flat(Infinity))];
-}
-
-function byString(obj, str) {
-    str = str.replace(/\[(\w+)\]/g, '.$1'); // convert indexes to properties
-    str = str.replace(/^\./, '');           // strip a leading dot
-    var a = str.split('.');
-    for (var i = 0, n = a.length; i < n; ++i) {
-        var k = a[i];
-        if (k in obj) 
-            obj = obj[k];
-        else return;
-    }
-    return obj;
-}
-
-/**
- * 
- * @param {string} in_camelCaseString 
- */
-function camelCaseToTitleCase(in_camelCaseString) {
-    var result = in_camelCaseString                         // "ToGetYourGEDInTimeASongAboutThe26ABCsIsOfTheEssenceButAPersonalIDCardForUser456InRoom26AContainingABC26TimesIsNotAsEasyAs123ForC3POOrR2D2Or2R2D"
-        .replace(/([a-z])([A-Z][a-z])/g, "$1 $2")           // "To Get YourGEDIn TimeASong About The26ABCs IsOf The Essence ButAPersonalIDCard For User456In Room26AContainingABC26Times IsNot AsEasy As123ForC3POOrR2D2Or2R2D"
-        .replace(/([A-Z][a-z])([A-Z])/g, "$1 $2")           // "To Get YourGEDIn TimeASong About The26ABCs Is Of The Essence ButAPersonalIDCard For User456In Room26AContainingABC26Times Is Not As Easy As123ForC3POOr R2D2Or2R2D"
-        .replace(/([a-z])([A-Z]+[a-z])/g, "$1 $2")          // "To Get Your GEDIn Time ASong About The26ABCs Is Of The Essence But APersonal IDCard For User456In Room26AContainingABC26Times Is Not As Easy As123ForC3POOr R2D2Or2R2D"
-        .replace(/([A-Z]+)([A-Z][a-z][a-z])/g, "$1 $2")     // "To Get Your GEDIn Time A Song About The26ABCs Is Of The Essence But A Personal ID Card For User456In Room26A ContainingABC26Times Is Not As Easy As123ForC3POOr R2D2Or2R2D"
-        .replace(/([a-z]+)([A-Z0-9]+)/g, "$1 $2")           // "To Get Your GEDIn Time A Song About The 26ABCs Is Of The Essence But A Personal ID Card For User 456In Room 26A Containing ABC26Times Is Not As Easy As 123For C3POOr R2D2Or 2R2D"
-        
-        // Note: the next regex includes a special case to exclude plurals of acronyms, e.g. "ABCs"
-        .replace(/([A-Z]+)([A-Z][a-rt-z][a-z]*)/g, "$1 $2") // "To Get Your GED In Time A Song About The 26ABCs Is Of The Essence But A Personal ID Card For User 456In Room 26A Containing ABC26Times Is Not As Easy As 123For C3PO Or R2D2Or 2R2D"
-        .replace(/([0-9])([A-Z][a-z]+)/g, "$1 $2")          // "To Get Your GED In Time A Song About The 26ABCs Is Of The Essence But A Personal ID Card For User 456In Room 26A Containing ABC 26Times Is Not As Easy As 123For C3PO Or R2D2Or 2R2D"  
-
-        // Note: the next two regexes use {2,} instead of + to add space on phrases like Room26A and 26ABCs but not on phrases like R2D2 and C3PO"
-        .replace(/([A-Z]{2,})([0-9]{2,})/g, "$1 $2")        // "To Get Your GED In Time A Song About The 26ABCs Is Of The Essence But A Personal ID Card For User 456 In Room 26A Containing ABC 26 Times Is Not As Easy As 123 For C3PO Or R2D2 Or 2R2D"
-        .replace(/([0-9]{2,})([A-Z]{2,})/g, "$1 $2")        // "To Get Your GED In Time A Song About The 26 ABCs Is Of The Essence But A Personal ID Card For User 456 In Room 26A Containing ABC 26 Times Is Not As Easy As 123 For C3PO Or R2D2 Or 2R2D"
-        .trim();
-
-
-    // capitalize the first letter
-    return result.charAt(0).toUpperCase() + result.slice(1);
-}
 
 passport.use(new LocalStrategy(function(username, pass, done) {
     db.verifyUser(username, pass).then((verified) => {
@@ -160,8 +110,8 @@ router.get("/research/:researchId", asyncWrap(async (req, res) => {
         res.render("research", {
             title: config.serverName,
             researchObj: research,
-            checkKeys,
-            byString 
+            checkKeys: mainLib.checkKeys,
+            byString: mainLib.byString 
         });
     }
     else {
@@ -177,28 +127,28 @@ router.get("/research/:researchId", asyncWrap(async (req, res) => {
 router.get("/research/:researchId/export", asyncWrap(async (req, res) => {
     let research = await db.findResearchById(req.params.researchId);
     if(research.researchers.includes(req.user.name) || req.user.permission > 0) {
-        let keys = checkKeys(research.data);
+        let keys = mainLib.checkKeys(research.data);
         let researchData = req.query.q ? research.data.filter(x=>{
-            logger.debug(`Export: ${keys[0]} ${byString(x, keys[0])} = ${byString(x, keys[0]).toLowerCase().includes(req.query.q)}`);
-            return byString(x, keys[0]).toLowerCase().includes(req.query.q);
+            logger.debug(`Export: ${keys[0]} ${mainLib.byString(x, keys[0])} = ${mainLib.byString(x, keys[0]).toLowerCase().includes(req.query.q)}`);
+            return mainLib.byString(x, keys[0]).toLowerCase().includes(req.query.q);
         }) : research.data;
         logger.info(`Exporting ${research.name}: Query ${req.query.q} vs ${keys[0]} => ${researchData.length}/${research.data.length} entries`);
 
-        keys = checkKeys(researchData);
+        keys = mainLib.checkKeys(researchData);
 
         let wb = new excel.Workbook();
         let ws = wb.addWorksheet(research.name);
 
         let row = 2, col = 1;
         for(let key of keys) {
-            ws.cell(1, col).string(camelCaseToTitleCase(key));
+            ws.cell(1, col).string(mainLib.camelCaseToTitleCase(key));
             col++;
         }
 
         for(let val of researchData) { 
             col = 1;
             for(let key of keys) {
-                let keyVal = byString(val, key);
+                let keyVal = mainLib.byString(val, key);
                 if (keyVal !== undefined) {
                     if (typeof keyVal === "number") 
                         ws.cell(row, col).number(keyVal);
